Another large one, check full message
When logging out and closing the app, reopening the application won't
automatically log you in if "Remember Me" was selected initially.

Flow management: ImFree is the 'root' which spawns intents, all others
when navigating finish() so at any point the user can click the hardware
back button to return to ImFree.

Layout update for LoginActivity, Removed back button in WhosFree