Push notifications and login handling
Push notifications are ready to be activiated. We can use usernames as
channels.

Login is getting better and better as you can't mash the login button
anymore as it magically transforms into a progress bar. In addition,
there's a quick patch to clear all activities from history when logging
out.